requirements.txt
Django==1.10.2
django-cors-headers==1.2.2
django-filter==0.15.2
djangorestframework==3.4.7
pkg-resources==0.0.0
